+++ tests/routerUtils.test.d.ts
@@ -1,0 +1,2 @@
+export {};
+//# sourceMappingURL=routerUtils.test.d.ts.map